requirements.txt
Flask
flask-cors
google-generativeai
requests
